make firebase init function to a module
